Add 'Top Scorer' example.
Suggested spec was to find 'Top Ten best scorers', 'The players who have scored the most
points when they [have] a defender closest to them'. This script filters the dataset to
only include shots that were made, sorts them by the distance to the closest defender,
and returns the player making the shot for the ten closest shots. Trying to make it
'distinct', giving each player only once, gave completely different results, so I'll
come back to this later.