Add 'Post Shooters' example.
Suggested spec was to find 'Best Starter and Finisher - non-post players (shots 5ft+
away from the basket) ... for all this take shot-attempts to be atleast 100'. This
may merely have been intended as common instructions for the four specific suggestions
that followed, but I decided to start with just this.

I'm more confident that this is correct than the 'Top Shooters' or 'Top Defenders'
examples.